requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
elasticsearch[async]==8.12.0
python-dotenv==1.0.0
pydantic==2.5.3
scikit-learn==1.4.0
numpy==1.26.3
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
